pathwaycom/pathway-pipeline/requirements.txt
pathway
python-dotenv
